+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Master
 
+* Nothing yet!
+
+## 5.1.0
+
 #### Updated
 
 * Add response methods for `Observable<DownloadRequest>`.